nuplayer: Fix de-referencing NULL mSource
InstantiateDecoderAction is executed after performing reset.
By the time performReset call is over, mSource is cleared and
instantiateDecoderAction is executed on a NULL mSource - causing
crash. At this point of time, mSource is already reset and it is
not intended to be used in future, so, we can safely return.
Check for NULL to avoid crash.

Change-Id: I5b748fe1f67f4699d4c9ab87e299c42ca6ea6a28
CRs-Fixed: 902831