nuplayer: Fix de-referencing NULL mSource
InstantiateDecoderAction is executed after performing reset.
By the time performReset call is over, mSource is cleared and
instantiateDecoderAction is executed on a NULL mSource - causing
crash. At this point of time, mSource is already reset and it is
not intended to be used in future, so, we can safely return.
Check for NULL to avoid crash.

Change-Id: Ie40854ef6d6a953966caf8bd17f4f4bb55e09d17
CRs-Fixed: 902831